Doing okay
Need to get the signup working:
Fixed playing the music with the process of trying to play the music from youtube, then going into google drive. However, then decided to just get to the web browser servers.

I need to get the username thing working and the current username variable.